moving methods from helpers into coordinate_systems
